+++ test/endpoint.js
@@ -39,6 +39,7 @@ describe('clustering', function() {
   require('./clustering/k-means.js');
   require('./clustering/key-collision.js');
   require('./clustering/naive.js');
+  require('./clustering/sorted-neighborhood.js');
   require('./clustering/vp-tree.js');
 });