+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 * Adding the `clustering/key-collision` namespace.
 * Adding the `clustering/naive` namespace.
+* Adding the `clustering/sorted-neighborhood` namespace.
 * Adding the `clustering/vp-tree` namespace.
 * Adding the `metrics/distance/identity` namespace.
 * Adding the `metrics/distance/monge-elkan` namespace.